remove document count to batched structure
